add genres
