+++ src/js/popular.js
@@ -1,6 +1,4 @@
-import axios from 'axios';
-
-import { markupFilm } from './markup';
+import { markupFilm } from './renderFilm';
 import { fetchResults } from './apiService';
 
 fetchResults(1).then(res => {
